fix(web): address pre-test PR review feedback
